Imported NRT_Loudness settings & updated BTTask
The audio synth now contains dedicated settings for the loudness factor that has been adjusted to begin reacting to audio when it is much louder than some tracks.

In the music currently used, secunda and radiation storm are considered 'quiet', and generally return 0 on the normalized loudness output. Sexualizer returns values between .3-.6 throughout the song.

I believe this will be the ideal settings for beginning experimentations solely based on loudness, before exploring amplitude. However, the AI's aggressiveness is a bool based on if the loudness is above .5, which may need to be adjusted in response to these NRTLoudness settings.