Removed BTTask_MoveRandomLocation
Confirmed Loudness was determining when the AI should be aggressive. Thus, the BTTask formerly used to move the AI is being removed to further test the aggressive factor.

The next task is to adjust the point when 'IsAggressive' is set, and perhaps add a delay to prevent jarring changes in behaviour, but rather smooth them out so when the song only becomes quiet for a moment, the AI still moves aggressively unless it's been longer than X seconds since it was loud.