Created Pre-gameloop analysis
On the beginplay event, the average loudness and constants (where notes fall on an array of bands. Ie; high frequencies might be focused on one end of the array, where they would be higher values than the rest) are calculated.

This calculation will be used as the basis for initial AI logic, and will be adjusted throughout the song based on the music playback in game-time.