Adds ass implementation of attack, which should be changed
Within the BP_NPC blueprint, a fixed function (that checks if the overlapped object class is a projectile) determines if the actor should take damage.

This should be changed so a call to the interface some c++ code (the npc class) will trigger child functions to fire their 'onattack' function. not sure if this is possible, but ANYTHING else would be better than this specific implementation.